feat: Update PaymentCubit to track selected payment method
The `PaymentCubit` class in `payment_cubit.dart` now includes a static variable `selectedPaymentMethod` to track the index of the selected payment method. This variable is used in the `PaymentMethodsListView` widget to update the active index when a payment method is tapped.

Refactor the code to improve the tracking of the selected payment method and enhance the user experience during the checkout process.